message: ledger,ledger-grp & bank api completed(error message,update & delete id not found, name change[manual])

+++ app/Api/V1_0/Accounting/Ledgers/Controllers/LedgerController.php
@@ -9,6 +9,8 @@ use ERP\Api\V1_0\Support\BaseController;
 use ERP\Api\V1_0\Accounting\Ledgers\Processors\LedgerProcessor;
 use ERP\Core\Accounting\Ledgers\Persistables\LedgerPersistable;
 use ERP\Core\Support\Service\ContainerInterface;
+use ERP\Exceptions\ExceptionMessage;
+use ERP\Model\Accounting\Ledgers\LedgerModel;
 /**
  * @author Reema Patel<reema.p@siliconbrain.in>
  */
@@ -16,12 +18,12 @@ class LedgerController extends BaseController implements ContainerInterface
 {
 	/**
      * @var ledgerService
-     * @var Processor
+     * @var processor
      * @var request
      * @var ledgerPersistable
      */
 	private $ledgerService;
-	private $Processor;
+	private $processor;
 	private $request;
 	private $ledgerPersistable;	
 	
@@ -51,7 +53,7 @@ class LedgerController extends BaseController implements ContainerInterface
 		if($requestMethod == 'POST')
 		{
 			$processor = new LedgerProcessor();
-			$ledgerPersistable = new LedgerPersistable();		
+			$ledgerPersistable = new LedgerPersistable();
 			$ledgerPersistable = $processor->createPersistable($this->request);
 			if($ledgerPersistable[0][0]=='[')
 			{
@@ -117,23 +119,32 @@ class LedgerController extends BaseController implements ContainerInterface
 	public function update(Request $request,$ledgerId)
     {    
 		$this->request = $request;
-		$Processor = new LedgerProcessor();
+		$processor = new LedgerProcessor();
 		$ledgerPersistable = new LedgerPersistable();		
-		$ledgerService= new LedgerService();			
-		$ledgerPersistable = $Processor->createPersistableChange($this->request,$ledgerId);
-		//here two array and string is return at a time
-		if($ledgerPersistable=="204: No Content Found For Update")
+		$ledgerService= new LedgerService();		
+		$ledgerModel = new LedgerModel();
+		$result = $ledgerModel->getData($ledgerId);
+		
+		//get exception message
+		$exception = new ExceptionMessage();
+		$fileSizeArray = $exception->messageArrays();
+		if(strcmp($result,$fileSizeArray['404'])==0)
 		{
-			return $ledgerPersistable;
-		}
-		else if(is_array($ledgerPersistable))
-		{
-			$status = $ledgerService->update($ledgerPersistable);
-			return $status;
+			return $result;
 		}
 		else
 		{
-			return $ledgerPersistable;
+			$ledgerPersistable = $processor->createPersistableChange($this->request,$ledgerId);
+			//here two array and string is return at a time
+			if(is_array($ledgerPersistable))
+			{
+				$status = $ledgerService->update($ledgerPersistable);
+				return $status;
+			}
+			else
+			{
+				return $ledgerPersistable;
+			}
 		}
 	}
 	
@@ -147,10 +158,23 @@ class LedgerController extends BaseController implements ContainerInterface
         $this->request = $request;
 		$Processor = new LedgerProcessor();
 		$ledgerPersistable = new LedgerPersistable();		
-		$ledgerService= new LedgerService();			
-		$ledgerPersistable = $Processor->createPersistableChange($this->request,$ledgerId);
-		$ledgerService->create($ledgerPersistable);
-		$status = $ledgerService->delete($ledgerPersistable);
-		return $status;
+		$ledgerService= new LedgerService();	
+		$ledgerModel = new LedgerModel();
+		$result = $ledgerModel->getData($ledgerId);
+		
+		//get exception message
+		$exception = new ExceptionMessage();
+		$fileSizeArray = $exception->messageArrays();
+		if(strcmp($result,$fileSizeArray['404'])==0)
+		{
+			return $result;
+		}
+		else
+		{		
+			$ledgerPersistable = $Processor->createPersistableChange($this->request,$ledgerId);
+			$ledgerService->create($ledgerPersistable);
+			$status = $ledgerService->delete($ledgerPersistable);
+			return $status;
+		}
     }
 }
